requirements.txt
flask
openai
python-dotenv
google-search-results